Add initial PyDrop application with drop probability calculator
This commit adds the initial version of PyDrop, a small Python application that allows to calculate the probability of obtaining an item when killing a monster in the online game Metin2. The application includes a drop probability calculator that uses predefined lists to calculate the percentage probability of a drop based on the level difference between the player and the monster. The calculator takes into account the number of monsters to be killed, the initial probability of getting the item, and the player and monster levels. The application is fully functional and tested.